% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carryTest.R
\name{carryTest}
\alias{carryTest}
\title{Check conjoint data for carryover effects}
\usage{
carryTest(data, outcome, attributes, task)
}
\arguments{
\item{data}{A data.frame (in long format).}

\item{outcome}{The outcome variable, rating or choice, entered as a string.}

\item{attributes}{A character vector of vignette attributes/dimensions.}

\item{task}{The task- or contest-ID; should be a factor in the data, entered here as a string.}
}
\value{
A list of class 'carryTest'. Can be converted to tidy data.frame with as.data.frame().
}
\description{
\code{carryTest} checks data from conjoint experiments for carryover effects
(which can arise when respondents base their evaluation of one vignette on their evaluation of previous ones).
It estimates linear regression models including one vignette attribute/dimension at a time interacted with an
ID for the task number as predictors and then estimates an F-test to test for the joint significance of the
interaction terms. A rejected null indicates carryover effects are present in the case of a particular
vignette attribute (see also Hainmueller et al. 2014, Political Analysis, p. 22).
}
\details{
The result can be plotted (using plot()) or exported as a tidy data.frame
(using as.data.frame()).
}
\examples{
\dontrun{
carryTest(data=experimentdata,
attributes=c("gender","age","income","education"),
outcome="rating",
task="taskID")
}

}
