% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimRandoTest.R
\name{dimRandoTest}
\alias{dimRandoTest}
\title{Checks if vignette attributes are uncorrelated}
\usage{
dimRandoTest(data, dims, labels = NULL)
}
\arguments{
\item{data}{A data.frame.}

\item{dims}{A character vector of vignette attributes' names in the dataset.}

\item{labels}{Optional. A character vector of labels for the vignette attributes
(Important: the order of the labels must correspond to the order of attributes in dims!).}
}
\value{
A list of class 'dimRandoTest'. Can be converted to a data.frame with as.data.frame().
}
\description{
\code{dimRandoTest} checks data from conjoint experiments for randomization
problems by checking for associations between any pair of profile attributes.
it cross-tabulates all combinations of profile attributes and computes chi-squared
tests to test if there are any significant associations between these attributes.
A rejected null (small p-value) indicates that there is some statistically
significant association between profile attributes. A Cramer's V test statistic
is also provided for each attribute combination to indicate the strength of any
association between the attributes. The output can be labeled.
}
\examples{
\dontrun{
dimRandoTest(data=experimentdata,
dims=c("age","nationality","gender"),
labels=c("Age","Nationality","Gender"))
}

}
