% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.carryTest.R
\name{plot.carryTest}
\alias{plot.carryTest}
\title{Produces simple bar plot of carryTest() result}
\usage{
\method{plot}{carryTest}(x, lcol = NULL, ltype = NULL, siglev = NULL, showsig = NULL, ...)
}
\arguments{
\item{x}{carryTest object}

\item{lcol}{Color of sign. threshold line (corresponds to 'col' argument for abline()).}

\item{ltype}{Type of sign. threshold line (corresponds to 'lty' argument for abline()).}

\item{siglev}{Significance threshold; the default is 0.05.}

\item{showsig}{Show vertical line indicating sign. threshold (default=TRUE)?}

\item{...}{ignored}
}
\description{
Produces simple bar plot of carryTest() result
}
\examples{
\dontrun{
result <- carryTest(data=experimentdata, attributes=c("gender","age","income",
"education"),outcome="rating",task="taskID")

plot(result, lcol = "red", ltype=5)
}

}
