% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.carryTest.R
\name{plot.carryTest}
\alias{plot.carryTest}
\title{Produces simple bar plot of carryTest() result}
\usage{
\method{plot}{carryTest}(
  x,
  showsig = NULL,
  lcol = NULL,
  ltype = NULL,
  siglev = NULL,
  margins = NULL,
  ...
)
}
\arguments{
\item{x}{carryTest object}

\item{showsig}{Show vertical line to indicate sign. threshold? (optional,
default=T)}

\item{lcol}{Color of sign. threshold line (optional, changes 'col' argument
for abline()).}

\item{ltype}{Type of sign. threshold line (optional, changes 'lty' argument
for abline()).}

\item{siglev}{Significance threshold (optional, the default is 0.05).}

\item{margins}{Plot margins (optional; changes par(mar())); must be a numeric
vector of length 4.}

\item{...}{ignored}
}
\description{
This creates a bar plot (using base R graphics) showing the main results
from a test for carryover effects: p-values per vignette attribute. For
custom (and publication-quality) graphs, export result as data.frame with
as.data.frame().
}
\examples{
\dontrun{
result <- carryTest(data=experimentdata, attributes=c("gender","age","income",
"education"),outcome="rating",task="taskID")

plot(result, lcol = "red", ltype=5)
}

}
