% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.carryTest.R
\name{plot.carryTest}
\alias{plot.carryTest}
\title{Produces simple bar plot of carryTest() result}
\usage{
\method{plot}{carryTest}(x, showsig = NULL, lcol = NULL, ltype = NULL, siglev = NULL, ...)
}
\arguments{
\item{x}{carryTest object}

\item{showsig}{Show vertical line indicating sign. threshold (default=TRUE)?}

\item{lcol}{Color of sign. threshold line (corresponds to 'col' argument
for abline()).}

\item{ltype}{Type of sign. threshold line (corresponds to 'lty' argument
for abline()).}

\item{siglev}{Significance threshold; the default is 0.05.}

\item{...}{ignored}
}
\description{
This creates a bar plot (using base R graphics) showing the main results
from a test for carryover effects: p-values per vignette attribute. For
custom (and publication-quality) graphs, export result as data.frame with
as.data.frame().
}
\examples{
\dontrun{
result <- carryTest(data=experimentdata, attributes=c("gender","age","income",
"education"),outcome="rating",task="taskID")

plot(result, lcol = "red", ltype=5)
}

}
