% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.dimRandoTest.R
\name{plot.dimRandoTest}
\alias{plot.dimRandoTest}
\title{Quick & easy option to plot result from dimRandoTest()}
\usage{
\method{plot}{dimRandoTest}(x, showsig = NULL, lcol = NULL, ltype = NULL, siglev = NULL, ...)
}
\arguments{
\item{x}{dimRandoTest object}

\item{showsig}{Show vertical line to indicate sign. threshold? (default=T)}

\item{lcol}{Color of sign. threshold line (corresponds to 'col' argument for abline()).}

\item{ltype}{Type of sign. threshold line (corresponds to 'lty' argument for abline()).}

\item{siglev}{Significance threshold; the default is 0.05.}

\item{...}{ignored.}
}
\description{
This creates a bar plot (using base R graphics) showing the main results
from a test for successful vignette attribute randomization. Shown are
p-values per vignette attribute based on chi-squared tests or one-way
ANOVA estimations (see ?dimRandoTest for details). For custom (and
publication-quality) graphs, export result as data.frame with as.data.frame().
}
\examples{
\dontrun{
result <- dimRandoTest(data=experimentdata,
attributes=c("age","nationality","gender"),
resvar = "respondent_gender",
vartype = "categorical")

plot(result)
}

}
