% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.dimRandoTest.R
\name{plot.dimRandoTest}
\alias{plot.dimRandoTest}
\title{Quick & easy option to plot result from dimRandoTest() as heatmap}
\usage{
\method{plot}{dimRandoTest}(x, margins = NULL, cexRow = NULL, cexCol = NULL, col = NULL, ...)
}
\arguments{
\item{x}{dimRandoTest object}

\item{margins}{Width of margins for labels. As margins parameter for heatmap().}

\item{cexRow}{Size of row labels. As cexRow parameter for heatmap().}

\item{cexCol}{Size of column labels. As cexCol parameter for heatmap().}

\item{col}{Color scheme. As col parameter for heatmap().}

\item{...}{ignored.}
}
\description{
Quick & easy option to plot result from dimRandoTest() as heatmap
}
\examples{
\dontrun{
result <- dimRandoTest(data=experimentdata,
dims=c("age","nationality","gender"),
labels=c("Age","Nationality","Gender"))

plot(result)
}

}
